requirements/dev.txt
black==22.3.0
km3net-testdata>=0.4.13
mock
numpydoc
pillow
pydocstyle
pytest
pytest-cov
pytest-flake8
pytest-pylint
pytest-watch
sphinx-autoapi
sphinx-gallery==0.10.1
sphinx-rtd-theme>=0.3
sphinx>=1.6.3
sphinx_rtd_theme>=0.3
sphinxcontrib-napoleon>=0.6.1
sphinxcontrib-programoutput>=0.11
sphinxcontrib-websupport>=1.0.1
twine
utm
wheel
